Strip markup out of USPS response and replace it with Unicode symbols.
Also, fix bug in RateV4 code.  Evidently, USPS assumes all times are in a particular timezone which is not Pacific time (where I am).  Testing at 11pm, the ship date timestamp was March 3 but the USPS web server rejected the transaction because the ship date has to be 0-3 days from the present day. From this I assumed that the USPS web server was at least one time zone east of Pacific time where it is now March 4.  Don't know exactly what it should be, but I'm changing it to Eastern time in hopes that is accurate.  (This is NOT documented in the API documentation).