Issue #1300608: Make sure tax line items are saved permanently on completed orders.
This fix involves separating the line items stored in the database from the ones
that have a value callback. Tax line items are both and need to be treated as
such. Thus we need two different tax line items. One to calculate the tax aomunt
on the fly, and one to store this amount.

The tax_display line item type plays more games with the line item weights. The
tax_subtotal line item needs to be calculated after tax_display, but actually
shown before it in the order. This is why uc_taxes_uc_line_item() gives
tax_display one weight, but its callback specifies another (the same one as the
stored tax line item, in fact).